requirements-backend.txt
aiogram==2.22.1
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
beanie==1.11.10
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
ecdsa==0.18.0
fastapi==0.85.0
frozenlist==1.3.1
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
kombu==5.2.4
motor==3.0.0
multidict==6.0.2
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
py==1.11.0
pyasn1==0.4.8
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typer==0.6.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
vine==5.0.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
yarl==1.8.1
